make models/teams.py and store team function in it
