+++ api/routes/v0/team.py
@@ -4,10 +4,9 @@ from helper.util import sha1_hash
 import datetime
 
 from helper.response import API_OK
-
 from models.requests import Team, TeamSimpleResponse
-
 from models.projects import Project
+from models.teams import Teams
 
 
 router = APIRouter()
@@ -15,7 +14,7 @@ router = APIRouter()
 @router.get("/", response_model=list[Team])
 def get_teams():
     try:
-        data = Project.get_teams()
+        data = Teams.get_teams()
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to get teams")
     return data
@@ -24,7 +23,7 @@ def get_teams():
 def post_team(team: Team):
     print("team")
     try:
-        res = Project.add_team(team)
+        res = Teams.add_team(team)
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to post team")
     print(res)
@@ -32,40 +31,40 @@ def post_team(team: Team):
 
 @router.post("/{team_id}/name", response_model=API_OK)
 def post_team_name(team_id: str, name: str):
-    if Project.get_team_by_id(team_id) is None:
+    if Teams.get_team_by_id(team_id) is None:
         raise StarletteHTTPException(status_code=404, detail="Team not found")
     try:
-        Project.update_team_name(team_id, name)
+        Teams.update_team_name(team_id, name)
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to post team name")
     return API_OK()
 
 @router.post("/{team_id}/year", response_model=API_OK)
 def post_team_year(team_id: str, year: int):
-    if Project.get_team_by_id(team_id) is None:
+    if Teams.get_team_by_id(team_id) is None:
         raise StarletteHTTPException(status_code=404, detail="Team not found")
     try:
-        Project.update_team_year(team_id, year)
+        Teams.update_team_year(team_id, year)
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to post team year")
     return API_OK()
 
 @router.post("/{team_id}/description", response_model=API_OK)
 def post_team_description(team_id: str, description: str):
-    if Project.get_team_by_id(team_id) is None:
+    if Teams.get_team_by_id(team_id) is None:
         raise StarletteHTTPException(status_code=404, detail="Team not found")
     try:
-        Project.update_team_description(team_id, description)
+        Teams.update_team_description(team_id, description)
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to post team description")
     return API_OK()
 
 @router.post("/{team_id}/members", response_model=API_OK)
 def post_team_members(team_id: str, new_members: list[str]):
-    if Project.get_team_by_id(team_id) is None:
+    if Teams.get_team_by_id(team_id) is None:
         raise StarletteHTTPException(status_code=404, detail="Team not found")
     try:
-        Project.add_team_members(team_id, new_members)
+        Teams.add_team_members(team_id, new_members)
     except:
         raise StarletteHTTPException(status_code=500, detail="Failed to post team members")
     return API_OK()
